feat: migrate from Create React App to Vite, update dependencies, and refactor actualities data handling
- Removed react-scripts and added Vite as the build tool.
- Updated package.json scripts for Vite commands.
- Added devDependencies for Vite and Vitest.
- Refactored actualities data imports to use raw markdown and updated image imports.
- Changed colleagues data import path to use assets directory.
- Created index.jsx as the entry point for the application.
- Added vite.config.mjs for Vite configuration.
