+++ index.html
@@ -12,10 +12,9 @@
     <link rel="manifest" href="/manifest.json" />
     <title>CSPM</title>
     <link rel="canonical" href="https://csaladodpenzugyimentora.hu/" />
-    <script type="module" crossorigin src="/assets/index-DnxKxiSs.js"></script>
-    <link rel="stylesheet" crossorigin href="/assets/index-BDvN2-1Y.css">
   </head>
   <body>
     <div id="root"></div>
+    <script type="module" src="/src/index.jsx"></script>
   </body>
 </html>
