+++ package.json
@@ -14,24 +14,25 @@
     "react-dom": "^18.3.1",
     "react-helmet": "^6.1.0",
     "react-markdown": "^9.0.1",
-    "react-scripts": "5.0.1",
     "react-slick": "^0.30.2",
     "react-swipeable": "^7.0.2",
     "slick-carousel": "^1.8.1",
     "web-vitals": "^2.1.4"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "release": "npm run build && npm version patch && git push origin main --follow-tags"
+  "devDependencies": {
+    "@vitejs/plugin-react": "^6.0.1",
+    "vite": "^8.0.10",
+    "vitest": "^4.1.5"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+  "scripts": {
+    "start": "vite",
+    "start:ext": "vite --host 0.0.0.0 --strictPort",
+    "start-ext": "vite --host 0.0.0.0 --strictPort",
+    "start:ext:open": "vite --host 0.0.0.0 --strictPort --open",
+    "build": "vite build",
+    "preview": "vite preview",
+    "test": "vitest run --passWithNoTests",
+    "release": "npm run build && npm version patch && git push origin main --follow-tags"
   },
   "browserslist": {
     "production": [
