+++ src/assets/actualitiesData.js
@@ -1,16 +1,22 @@
-import allampapirokContent from './actualities/allampapirok.md';
-import onypContent from './actualities/onyp.md';
-import szkolcsonContent from './actualities/szemelyi_kolcson.md';
-import penzugyKarrierLehetosegContent from './actualities/penzugy_karrier_hirdetes.md';
-import udvContent from './actualities/udv.md';
-import healthCare00Content from './actualities/healthcare00.md';
+import allampapirokContent from './actualities/allampapirok.md?raw';
+import onypContent from './actualities/onyp.md?raw';
+import szkolcsonContent from './actualities/szemelyi_kolcson.md?raw';
+import penzugyKarrierLehetosegContent from './actualities/penzugy_karrier_hirdetes.md?raw';
+import udvContent from './actualities/udv.md?raw';
+import healthCare00Content from './actualities/healthcare00.md?raw';
+import karrierLehetosegImage from './adbar/karrier_lehetoseg.png';
+import szemelyiKolcsonImage from './adbar/szemelyi_kolcson.png';
+import cspmFbAd01Image from './adbar/cspm_fb_ad_01.png';
+import allampapirokImage from './adbar/allampapirok.png';
+import healthcare00Image from './adbar/healthcare00.png';
+import udvImage from './adbar/udv.png';
 
 // Define your actuality data
 const actualitesData = [
   {
     id: 1,
     title: 'Karrier lehetőség',
-    image: require('./adbar/karrier_lehetoseg.png'),
+    image: karrierLehetosegImage,
     contentPath: penzugyKarrierLehetosegContent,
     summary: 'Készen állsz a következő lépésre a karrieredben? Csatlakozz csapatunkhoz!'
 
@@ -18,35 +24,35 @@ const actualitesData = [
   {
     id: 2,    
     title: 'Személyi kölcsön',
-    image: require('./adbar/szemelyi_kolcson.png'),
+    image: szemelyiKolcsonImage,
     contentPath: szkolcsonContent,
     summary: 'Nagyszerű terveid vannak? Segítünk megvalósítani az álmaidat!'
 },
   {
     id: 3,
     title: 'Nyugdíjpénztár felhasználása lakáscélokra',
-    image: require('./adbar/cspm_fb_ad_01.png'),
+    image: cspmFbAd01Image,
     contentPath: onypContent,
     summary: 'Új lehetőség az önkéntes nyugdíjpénztári megtakarítások lakáscélú felhasználására'
   },
   {
     id: 4,
     title: 'Állampapírok vagy Befektetés + Biztosítás',
-    image: require('./adbar/allampapirok.png'),
+    image: allampapirokImage,
     contentPath: allampapirokContent,
     summary: 'Melyik befektetési forma a legmegfelelőbb az Ön számára? Ismerje meg a lehetőségeket!'
   },
   {
     id: 5,
     title: 'Új Baleseti Biztosítási termékünk!',
-    image: require('./adbar/healthcare00.png'),
+    image: healthcare00Image,
     contentPath: healthCare00Content,
     summary: 'Egy rossz lépés ne kerüljön sokba! Baleseti sérülésekre szóló kiegészítő, amely tényleg térít! 🩹'
   },
     {
     id: 6,
     title: 'Üdvözlöm a CsPM honlapján!',
-    image: require('./adbar/udv.png'),
+      image: udvImage,
     contentPath: udvContent,
     summary: 'Kollégáim és én azért dolgozunk, hogy Önnek a legjobb pénzügyi megoldásokat kínáljuk. Célunk, hogy segítsünk Önnek a pénzügyei kezelésében és a jövője biztosításában.'
   },
